add improved patter matcher and test framework
The new **pattern matcher** is the third approximation to ELIZA and it
follows better the idioms of scheme, in particular: the atom? predicate
is not used, consistent order of arguments in procedure calls and
consistent naming conventions.

A **test framework** was introduced to make tests easy to write, this
framework is a modified and simplified version of Neil Van Dyke's
Testeez *obsolete* code.